% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.seqs.R
\name{find.seqs}
\alias{find.seqs}
\title{Find out how many nucleotide records are in NCBIs database}
\usage{
find.seqs(Species = c("Atheresthes evermanni", "Atheresthes stomias"),
  Loci = c("Cytb", "cytochrome b", "dloop", "d loop", "control region",
  "CR", "COI", "COXI", "CO1", "COII", "COIII", "16S", "12S", "ND5", "ND3"),
  LengthRange = "100:20000", LociGrps = list(1:2, 3:6, 7:9, 10, 11, 12,
  13, 14, 15), ...)
}
\arguments{
\item{Species}{A vector of species names.}

\item{Loci}{A vector of loci to search.}

\item{LociGrps}{A list of how to group the loci if you want to search multiple names for the same locus (ie. COI, COXI, cytochrome oxidase I)}

\item{Length}{range for sequences to keep default is 100:20000}
}
\description{
This function facilitates the search of NCBIs nucleotide db to find sequences
}
\examples{
find.seqs()
}
\keyword{NCBI}
\keyword{nucleotide}
