% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find.SpcsSynyms.R
\name{Find.SpcsSynyms}
\alias{Find.SpcsSynyms}
\title{Find species synonyms}
\usage{
Find.SpcsSynyms(Species, ...)
}
\arguments{
\item{Species}{A vector of species names.}
}
\description{
This function facilitates the search of NCBIs taxonomy db to find species synonyms
}
\examples{
Find.SpcsSynyms()
}
\keyword{NCBI}
\keyword{taxonomy}
