% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.synyms.R
\name{find.synyms}
\alias{find.synyms}
\title{Find species synonyms}
\usage{
find.synyms(Species, ...)
}
\arguments{
\item{Species}{A vector of species names.}
}
\description{
This function facilitates the search of NCBIs taxonomy db to find species synonyms
}
\examples{
find.synyms()
}
\keyword{NCBI}
\keyword{species}
\keyword{synonyms}
\keyword{taxonomy}
