% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.Lineage.R
\name{Get.lineage}
\alias{Get.lineage}
\title{Get lineage}
\usage{
Get.lineage(Species, NCBI_TaxaIDs, GetBy = "species", ...)
}
\arguments{
\item{Species}{A character vector of species names.}

\item{NCBI_TaxaIDs}{A character vector of taxa IDs}

\item{GetBy}{either "species" or "taxaID"}
}
\description{
This function facilitates getting the taxonomic lineage of a species
}
\examples{
Get.lineage()
}
\keyword{NCBI}
\keyword{nucleotide}
