% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.taxonomy.R
\name{get.taxonomy}
\alias{get.taxonomy}
\title{Get taxonomy}
\usage{
get.taxonomy(Species, NCBI_TaxaIDs, GetBy = "species", ...)
}
\arguments{
\item{Species}{A character vector of species names.}

\item{NCBI_TaxaIDs}{A character vector of taxa IDs}

\item{GetBy}{either "species" or "taxaID"}
}
\description{
This function facilitates getting the taxonomic lineage of a species
}
\examples{
get.taxonomy()
}
\keyword{NCBI}
\keyword{nucleotide}
