% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.seqs.R
\name{pull.seqs}
\alias{pull.seqs}
\title{Pull nucleotide sequences from NCBI nucleotide database}
\usage{
\method{pull}{seqs}(Species, Loci, LengthRange = "100:20000", LociGrps,
  ...)
}
\arguments{
\item{Species}{A vector of species names.}

\item{Loci}{A vector of loci to search.}

\item{LociGrps}{A list of how to group the loci if you want to search multiple names for the same locus (ie. COI, COXI, cytochrome oxidase I)}

\item{Length}{range for sequences to keep default is 100:20000}
}
\description{
This function facilitates the download of NCBIs nucleotide db
}
\examples{
pull.seqs()
}
\keyword{NCBI}
\keyword{nucleotide}
